Refactor: extract ProgressHelper into its own header
Move ProgressHelper class from rdfutils.hpp to a new dedicated
progresshelper.hpp, and add #pragma once to rdfutils.hpp. This
reduces the size and scope of rdfutils.hpp.
